+++ src/Farmer/Arm/ContainerService.fs
@@ -3,6 +3,7 @@ module Farmer.Arm.ContainerService
 
 open Farmer
 open Farmer.CoreTypes
+open Farmer.Identity
 open Farmer.Vm
 
 let managedClusters = ResourceType ("Microsoft.ContainerService/managedClusters", "2020-04-01")
@@ -25,6 +26,7 @@ type ManagedCluster =
         |} list
       DnsPrefix : string
       EnableRBAC : bool
+      Identity : ManagedIdentity
       LinuxProfile :
        {| AdminUserName : string
           PublicKeys : string list |} option
@@ -54,12 +56,15 @@ type ManagedCluster =
     interface IArmResource with
         member this.ResourceName = this.Name
         member this.JsonModel =
-            let dependencies =
-               this.AgentPoolProfiles
-               |> List.map (fun pool -> pool.VirtualNetworkName)
-               |> List.choose id
-               |> List.map(fun vnet -> ResourceId.create(virtualNetworks, vnet))
+            let dependencies = [
+                   yield!
+                       this.AgentPoolProfiles
+                       |> List.choose (fun pool -> pool.VirtualNetworkName)
+                       |> List.map(fun vnet -> ResourceId.create(virtualNetworks, vnet))
+                   yield! this.Identity.Dependencies
+               ]
             {| managedClusters.Create(this.Name, this.Location, dependencies) with
+                   identity = this.Identity |> ManagedIdentity.toArmJson
                    properties =
                        {| agentPoolProfiles =
                            this.AgentPoolProfiles