+++ src/Tests/AllTests.fs
@@ -38,6 +38,7 @@ let allTests =
             ]
             testList "Control" [
                 Template.tests
+                Identity.tests
                 Common.tests
                 if Environment.GetEnvironmentVariable "TF_BUILD" = "True" then AzCli.tests
             ]