+++ RELEASE_NOTES.md
@@ -1,12 +1,19 @@
 Release Notes
 =============
 ## 1.2.0-beta2
+* Managed Identity: Create user assigned manage identity resources.
+* Container Service (AKS): Support for Managed Identity.
+* Container Groups: Support for Managed Identity.
 * Container Groups: Support for image registry credentials for private registries.
 * Container Groups: Support for partial CPU cores.
 * Log Analytics: Initial support.
+* Web App: Full support for Managed Identity (minor breaking change).
+* Key Vault: Grant access to managed identities.
+* Storage Account: Grant access to managed identities.
 * Static Web Apps: Initial support.
 * Storage Accounts: Upgrade API version to 2019-04-01 to support RA-GZRS.
 * Internal changes to better capture full resource IDs.
+* Provide all Roles for managed identity purposes.
 * Event Hubs: Remove redundant kafka flag (breaking change).
 
 ## 1.2.0-beta1