+++ RELEASE_NOTES.md
@@ -1,13 +1,11 @@
 Release Notes
 =============
-## 1.2.0
-* Add branding.
-
 ## 1.2.0-beta2
 * Container Groups: Support for image registry credentials for private registries.
 * Container Groups: Support for partial CPU cores.
 * Log Analytics: Initial support.
 * Internal changes to better capture full resource IDs.
+* Add branding.
 
 ## 1.2.0-beta1
 * Bastion Hosts: Create bastion hosts for accessing resources on a virtual network